Rename PydanticAI ticket resolver example
