+++ .gitignore
@@ -295,4 +295,4 @@ replay-verify-reports/
 openapi/.build/
 
 # Generated by the canonical evaluator-authoring walkthrough
-/examples/canonical_example/evaluator.py
+/examples/pydantic_ai_ticket_resolver/evaluator.py
