+++ examples/pydantic_ai_ticket_resolver/.env.example
