+++ examples/pydantic_ai_ticket_resolver/__init__.py
