+++ examples/pydantic_ai_ticket_resolver/generate.sh
@@ -18,6 +18,6 @@ printf '%s\n' 'Generating 10 returns-resolution traces in Langfuse'
 uv run --env-file "${env_file}" \
   --extra examples \
   --with-editable "${repository_root}/plugins/packages/pydantic-ai[openai]" \
-  python -m examples.canonical_example.generate_traces "${trace_file}"
+  python -m examples.pydantic_ai_ticket_resolver.generate_traces "${trace_file}"
 
 printf 'Wrote %s\n' "${trace_file}"
