+++ examples/pydantic_ai_ticket_resolver/models.py
