+++ examples/pydantic_ai_ticket_resolver/store.py
@@ -2,8 +2,8 @@
 
 from decimal import Decimal
 
-from examples.canonical_example.fixtures import ORDERS, POLICIES, SHIPMENTS
-from examples.canonical_example.models import (
+from examples.pydantic_ai_ticket_resolver.fixtures import ORDERS, POLICIES, SHIPMENTS
+from examples.pydantic_ai_ticket_resolver.models import (
     ActionReceipt,
     Order,
     OrderLookup,
